feat: Code generator for AST and cli improvements
The functional implementation of the code generator for AST, in charge of creating the higher hierarchy class together with the lower hierarchy classes and their initializing functions, was implemented. Important improvements were implemented in the aspect of alert reports in the console as well as the arguments handled.

New way to use the interactive console, as well as a new way to read code from a file, and a new command to manage the different tools for development.